permissions: Add scc and try to set minimal capabilities to run ipa-server-install
The following list of capabilities seems to be the required to run
ipa-server-install properly:

- "CHOWN"
- "FOWNER"
- "DAC_OVERRIDE"
- "SETUID"
- "SETGID"
- "SETFCAP"
- "SETPCAP"
- "KILL"
- "NET_BIND_SERVICE"
- "SYS_ADMIN"
- "SYS_RESOURCE"

A ServiceAccount, SecurityContextConstraints, Role, ClusterRole and the
corresponding bindings are provided on this proof of concept which could
be used as base objects for the operator.
This provide too a set of mountpoints needed to run the initContainer
properly and initialize the /data volume. More details can be found at
the README.md file, and a quick start guide for the proof of concept.